Refactored Dropdown component
