+++ app/javascript/src/components/Dashboard/Contacts/Table.jsx
@@ -1,18 +1,9 @@
 import React, { useState } from "react";
 
-import { MenuHorizontal } from "@bigbinary/neeto-icons";
-import {
-  Alert,
-  Avatar,
-  Dropdown,
-  Table as NeetoUITable,
-  Toastr,
-  Tooltip,
-  Typography,
-} from "@bigbinary/neetoui";
+import { Alert, Table as NeetoUITable, Toastr } from "@bigbinary/neetoui";
 import { useTranslation } from "react-i18next";
 
-import { contactsTableData } from "./utils";
+import { contactsTableColumnData, contactsTableData } from "./utils";
 
 const Table = () => {
   const [pageNumber, setPageNumber] = useState(1);
@@ -26,69 +17,18 @@ const Table = () => {
     Toastr.success(t("toasts.deleted_entity", { entity: "Contact" }));
   };
 
-  const COLUMN_DATA = [
-    {
-      title: t("contacts_table.title.name_and_role"),
-      dataIndex: "name",
-      key: "name",
-      width: "30%",
-      render: (_, contact) => (
-        <div className="flex gap-x-3">
-          <div>
-            <Avatar size="large" />
-          </div>
-          <div className="flex flex-col">
-            <Typography style="h5">{contact.name}</Typography>
-            <Typography style="body3">{contact.role}</Typography>
-          </div>
-        </div>
-      ),
-    },
-    {
-      title: t("contacts_table.title.email"),
-      dataIndex: "email",
-      key: "email",
-      width: "30%",
-    },
-    {
-      title: t("contacts_table.title.created_at"),
-      dataIndex: "createdAt",
-      key: "createdAt",
-      width: "30%",
-      render: createdAt => <Tooltip content={createdAt}>{createdAt}</Tooltip>,
-    },
-    {
-      title: "",
-      dataIndex: "menu",
-      key: "menu",
-      width: 40,
-      render: () => {
-        const { Menu, MenuItem } = Dropdown;
-
-        return (
-          <Dropdown buttonStyle="text" icon={MenuHorizontal}>
-            <Menu>
-              <MenuItem.Button>{t("edit")}</MenuItem.Button>
-              <MenuItem.Button onClick={() => setIsDeleteAlertOpen(true)}>
-                {t("delete")}
-              </MenuItem.Button>
-            </Menu>
-          </Dropdown>
-        );
-      },
-    },
-  ];
-
   return (
     <>
       <NeetoUITable
         rowSelection
-        columnData={COLUMN_DATA}
         currentPageNumber={pageNumber}
         defaultPageSize={9}
         handlePageChange={page => setPageNumber(page)}
         rowData={contactsTableData()}
         selectedRowKeys={selectedRowKeys}
+        columnData={contactsTableColumnData({
+          onContactDelete: () => setIsDeleteAlertOpen(true),
+        })}
         onRowSelect={selectedRowKeys => setSelectedRowKeys(selectedRowKeys)}
       />
       <Alert
