+++ app/javascript/src/components/Dashboard/Notes/Item.jsx
@@ -1,15 +1,10 @@
 import React, { useState } from "react";
 
-import { Clock, MenuVertical } from "@bigbinary/neeto-icons";
-import {
-  Alert,
-  Avatar,
-  Dropdown,
-  Tag,
-  Tooltip,
-  Typography,
-} from "@bigbinary/neetoui";
+import { Clock } from "@bigbinary/neeto-icons";
+import { Alert, Avatar, Tag, Tooltip, Typography } from "@bigbinary/neetoui";
 import { useTranslation } from "react-i18next";
+
+import Dropdown from "components/commons/Dropdown";
 
 import { calculateElapsedTime, formatTimeForTooltip } from "./utils";
 
@@ -17,24 +12,15 @@ const Item = ({ note, onDelete }) => {
   const [isDeleteAlertOpen, setIsDeleteAlertOpen] = useState(false);
   const { t } = useTranslation();
 
-  const { Menu, MenuItem } = Dropdown;
-
   return (
     <>
       <div className="flex w-full flex-col gap-y-2 rounded-sm border border-solid p-4 shadow-sm">
         <div className="flex justify-between">
           <Typography style="h4">{note.title}</Typography>
           <Dropdown
-            buttonStyle="text"
-            icon={() => <MenuVertical size="15px" />}
-          >
-            <Menu>
-              <MenuItem.Button>{t("edit")}</MenuItem.Button>
-              <MenuItem.Button onClick={() => setIsDeleteAlertOpen(true)}>
-                {t("delete")}
-              </MenuItem.Button>
-            </Menu>
-          </Dropdown>
+            iconType="vertical"
+            onDelete={() => setIsDeleteAlertOpen(true)}
+          />
         </div>
         <Typography className="neeto-ui-text-gray-600" style="body2">
           {note.description}
