+++ app/javascript/src/components/commons/constants.js
@@ -1,0 +1,2 @@
+export const DROPDOWN_ICON_TYPES = ["vertical", "horizontal"];
+export const DROPDOWN_DEFAULT_ICON_TYPE = "vertical";
